trim the movePlayer function
This fixs the former promblem. I use the main class to connect player
and dice. So, when create dice, I pass a main(this) into it as
documentClass. Then, we can use this documentClass to call a movePlayer
function which is in main.class, and use it to transfer dicePoints.
Moveplayer function check the rolling status of both players and call
function update the position.