Snake&Ladder functions work
currently, use the basic way put snakes and ladders into fixed
positions, and according their position add the top and bottom tile
number into a dictionary: The start point is as the key, and end point
is as the value. And add a check function in player class and use
current position to find the next position, if player can climb the
ladder or is catched by the snake.